refactor: Url을 application 영역의 패키지로 이동
