+++ tomcat/src/test/java/nextstep/jwp/UrlTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package nextstep.jwp;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,6 +7,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import nextstep.jwp.exception.NotFoundException;
+import org.apache.coyote.http11.Http11Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
